pki.desciptor: moved String() methods from implementation to main.
